+++ react/SellingPrice.tsx
@@ -6,6 +6,7 @@ import { IOMessageWithMarkers } from 'vtex.native-types'
 import { useCssHandles, CssHandlesTypes } from 'vtex.css-handles'
 
 import { getDefaultSeller } from './modules/seller'
+import { hideProductPrice } from './modules/hideProductPrice'
 
 const CSS_HANDLES = [
   'sellingPrice',
@@ -35,12 +36,14 @@ interface Props {
   markers?: string[]
   /** Used to override default CSS handles */
   classes?: CssHandlesTypes.CustomClasses<typeof CSS_HANDLES>
+  alwaysShow?: boolean
 }
 
 function SellingPrice({
   message = messages.default.id,
   markers = [],
   classes,
+  alwaysShow = false,
 }: Props) {
   const { handles, withModifiers } = useCssHandles(CSS_HANDLES, {
     classes,
@@ -52,7 +55,13 @@ function SellingPrice({
 
   const commercialOffer = seller?.commertialOffer
 
-  if (!commercialOffer || commercialOffer?.AvailableQuantity <= 0) {
+  if (
+    !commercialOffer ||
+    hideProductPrice({
+      alwaysShow,
+      availableQuantity: commercialOffer.AvailableQuantity,
+    })
+  ) {
     return null
   }
 
@@ -78,6 +87,7 @@ function SellingPrice({
     hasListPrice ? 'hasListPrice' : '',
     hasMeasurementUnit ? 'hasMeasurementUnit' : '',
     hasUnitMultiplier ? 'hasUnitMultiplier' : '',
+    alwaysShow && commercialOffer.AvailableQuantity <= 0 ? 'isUnavailable' : '',
   ])
 
   return (