+++ react/SellingPriceRange.tsx
@@ -6,6 +6,7 @@ import { useCssHandles, CssHandlesTypes } from 'vtex.css-handles'
 import { IOMessageWithMarkers } from 'vtex.native-types'
 
 import { getDefaultSeller } from './modules/seller'
+import { hideProductPrice } from './modules/hideProductPrice'
 
 const CSS_HANDLES = [
   'sellingPriceRange',
@@ -49,6 +50,7 @@ interface Props {
   markers?: string[]
   /** Used to override default CSS handles */
   classes?: CssHandlesTypes.CustomClasses<typeof CSS_HANDLES>
+  alwaysShow?: boolean
 }
 
 function SellingPriceRange({
@@ -56,8 +58,9 @@ function SellingPriceRange({
   noRangeMessage = messages.noRangeMessageDefault.id,
   markers = [],
   classes,
+  alwaysShow = false,
 }: Props) {
-  const { handles } = useCssHandles(CSS_HANDLES, { classes })
+  const { handles, withModifiers } = useCssHandles(CSS_HANDLES, { classes })
   const productContextValue = useProduct()
 
   const priceRange = productContextValue?.product?.priceRange
@@ -70,7 +73,13 @@ function SellingPriceRange({
 
   const commercialOffer = seller?.commertialOffer
 
-  if (!commercialOffer || commercialOffer?.AvailableQuantity <= 0) {
+  if (
+    !commercialOffer ||
+    hideProductPrice({
+      alwaysShow,
+      availableQuantity: commercialOffer.AvailableQuantity,
+    })
+  ) {
     return null
   }
 
@@ -84,9 +93,13 @@ function SellingPriceRange({
     commercialOffer.PriceWithoutDiscount +
     commercialOffer.PriceWithoutDiscount * commercialOffer.taxPercentage
 
+  const containerClasses = withModifiers('sellingPriceRange', [
+    alwaysShow && commercialOffer.AvailableQuantity <= 0 ? 'isUnavailable' : '',
+  ])
+
   if (hasRange) {
     return (
-      <span className={handles.sellingPriceRange}>
+      <span className={containerClasses}>
         <IOMessageWithMarkers
           message={message}
           markers={markers}
@@ -147,7 +160,7 @@ function SellingPriceRange({
   }
 
   return (
-    <span className={handles.sellingPriceRange}>
+    <span className={containerClasses}>
       <IOMessageWithMarkers
         message={noRangeMessage}
         markers={markers}